[NuGet] Add custom build of NuGet 3.5 with proxy and restore fix
Workaround for proxy credentials not being requested for https
package sources on Mono. Mono throws a different exception when a
request to a https url requires proxy authentication which is now
handled in NuGet.

Reduced the number of parallel tasks from 16 down to 1. Xamarin Studio
has a limited number of threadpool threads and the NuGet package
restore can cause them to be all used and then the restore gets stuck
and never ends.